nyxNOVA-be/requirements.txt
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
defusedxml==0.7.1
fastapi==0.114.0
filelock==3.15.4
filetype==1.2.0
Flask==3.0.3
Flask-Cors==5.0.0
fonttools==4.53.1
fsspec==2024.6.1
gevent==24.2.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
huggingface-hub==0.23.5
idna==3.7
intel-openmp==2021.4.0
itsdangerous==2.2.0
Jinja2==3.1.3
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
llvmlite==0.43.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
mdurl==0.1.2
mkl==2021.4.0
mpmath==1.3.0
networkx==3.2.1
numba==0.60.0
numpy==1.26.4
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.1
pandas==2.2.2
pillow==10.2.0
protobuf==5.27.3
psutil==6.0.0
py-cpuinfo==9.0.0
pyarrow==17.0.0
pycparser==2.22
pydantic==2.9.0
pydantic_core==2.23.2
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.7.1
roboflow==1.1.36
rpds-py==0.19.1
safetensors==0.4.3
scipy==1.14.0
seaborn==0.13.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
starlette==0.38.4
streamlit==1.37.0
supervision==0.22.0
sympy==1.12
tbb==2021.11.0
tenacity==8.5.0
thop==0.1.1.post2209072238
toml==0.10.2
toolz==0.12.1
torch==2.3.1+cu118
torchvision==0.18.1+cu118
tornado==6.4.1
tqdm==4.66.4
transformers==4.42.4
typing_extensions==4.12.2
tzdata==2024.1
ultralytics @ git+https://github.com/THU-MIG/yolov10.git@aad320dd80b56694e590c950b25060a134966496
urllib3==2.2.2
watchdog==4.0.1
Werkzeug==3.0.4
zope.event==5.0
zope.interface==7.0.3
